Explicit variable cast to have less compiler warnings(I hope it's safe :) )

+++ Sources/hoa.decoder_tilde.cpp
@@ -39,7 +39,7 @@ static void *hoa_decoder_new(t_symbol *s, long argc, t_atom *argv)
     t_symbol* mode = gensym("regular");
     t_hoa_decoder *x = (t_hoa_decoder *)eobj_new(hoa_decoder_class);
     t_binbuf *d      = binbuf_via_atoms(argc,argv);
-	
+
     if(x && d)
 	{
         if(argc && argv && atom_gettype(argv) == A_LONG)
@@ -48,7 +48,7 @@ static void *hoa_decoder_new(t_symbol *s, long argc, t_atom *argv)
             mode = atom_getsym(argv+1);
         if(argc > 2 && argv+2 && atom_gettype(argv+2) == A_LONG)
             number_of_channels = pd_clip_min(atom_getlong(argv+2), 1);
-        
+
         if(mode == gensym("irregular"))
         {
             x->f_decoder = new Decoder<Hoa2d, t_sample>::Irregular(order, number_of_channels);
@@ -71,10 +71,10 @@ static void *hoa_decoder_new(t_symbol *s, long argc, t_atom *argv)
         memset(x->f_ins, 0, x->f_decoder->getNumberOfHarmonics() * HOA_MAXBLKSIZE * sizeof(t_sample));
         memset(x->f_outs, 0, x->f_decoder->getNumberOfPlanewaves() * HOA_MAXBLKSIZE * sizeof(t_sample));
         ebox_attrprocess_viabinbuf(x, d);
-        
+
         return x;
 	}
-    
+
     return NULL;
 }
 
@@ -142,14 +142,14 @@ static t_pd_err hoa_decoder_angles_set(t_hoa_decoder *x, void *attr, long argc, 
     if(argc && argv)
     {
         int dspState = canvas_suspend_dsp();
-        for(int i = 0; i < argc && i < x->f_decoder->getNumberOfPlanewaves(); i++)
+        for(long i = 0; i < argc && i < (long)x->f_decoder->getNumberOfPlanewaves(); i++)
         {
             if(atom_gettype(argv+i) == A_FLOAT)
                 x->f_decoder->setPlanewaveAzimuth(i, atom_getfloat(argv+i) / 360. * HOA_2PI);
         }
         canvas_resume_dsp(dspState);
     }
-    
+
     return 0;
 }
 
@@ -210,26 +210,26 @@ static void hoa_decoder_free(t_hoa_decoder *x)
 extern "C" void setup_hoa0x2e2d0x2edecoder_tilde(void)
 {
     t_eclass* c;
-    
+
     c = eclass_new("hoa.2d.decoder~", (method)hoa_decoder_new, (method)hoa_decoder_free, (short)sizeof(t_hoa_decoder), 0L, A_GIMME, 0);
     class_addcreator((t_newmethod)hoa_decoder_new, gensym("hoa.decoder~"), A_GIMME, 0);
-    
+
     eclass_dspinit(c);
     hoa_initclass(c);
     eclass_addmethod(c, (method)hoa_decoder_dsp,           "dsp",          A_CANT,  0);
-    
+
     CLASS_ATTR_DOUBLE_VARSIZE	(c, "angles",0, t_hoa_decoder, f_attrs, f_attrs, HOA_MAX_PLANEWAVES);
     CLASS_ATTR_ACCESSORS		(c, "angles", hoa_decoder_angles_get, hoa_decoder_angles_set);
     CLASS_ATTR_CATEGORY			(c, "offset", 0, "Planewaves");
     CLASS_ATTR_LABEL			(c, "angles", 0, "Angles of Loudspeakers");
     CLASS_ATTR_SAVE             (c, "angles", 0);
-    
+
     CLASS_ATTR_DOUBLE           (c, "offset", 0, t_hoa_decoder, f_attrs);
     CLASS_ATTR_ACCESSORS		(c, "offset", hoa_decoder_offset_get, hoa_decoder_offset_set);
     CLASS_ATTR_CATEGORY			(c, "offset", 0, "Planewaves");
     CLASS_ATTR_LABEL            (c, "offset", 0, "Offset of Channels");
     CLASS_ATTR_SAVE             (c, "offset", 0);
-    
+
     eclass_register(CLASS_OBJ, c);
     hoa_decoder_class = c;
 }
@@ -241,7 +241,7 @@ static void *hoa_decoder_3d_new(t_symbol *s, long argc, t_atom *argv)
     int number_of_channels = 4;
     t_hoa_decoder_3d *x = (t_hoa_decoder_3d *)eobj_new(hoa_decoder_3d_class);
     t_binbuf *d         = binbuf_via_atoms(argc,argv);
-    
+
     if(x && d)
     {
         if(argc && argv && atom_gettype(argv) == A_LONG)
@@ -250,7 +250,7 @@ static void *hoa_decoder_3d_new(t_symbol *s, long argc, t_atom *argv)
             mode = atom_getsym(argv+1);
         if(argc > 2 && argv+2 && atom_gettype(argv+2) == A_LONG)
             number_of_channels = pd_clip_min(atom_getlong(argv+2), 1);
-        
+
         if(mode == gensym("irregular"))
         {
             x->f_decoder = new Decoder<Hoa3d, t_sample>::Regular(order, number_of_channels);
@@ -263,17 +263,17 @@ static void *hoa_decoder_3d_new(t_symbol *s, long argc, t_atom *argv)
         {
             x->f_decoder = new Decoder<Hoa3d, t_sample>::Regular(order, number_of_channels);
         }
-        
+
         eobj_dspsetup(x, x->f_decoder->getNumberOfHarmonics(), x->f_decoder->getNumberOfPlanewaves());
         x->f_ins = new t_sample[x->f_decoder->getNumberOfHarmonics() * HOA_MAXBLKSIZE];
         x->f_outs= new t_sample[x->f_decoder->getNumberOfPlanewaves() * HOA_MAXBLKSIZE];
         memset(x->f_ins, 0, x->f_decoder->getNumberOfHarmonics() * HOA_MAXBLKSIZE * sizeof(t_sample));
         memset(x->f_outs, 0, x->f_decoder->getNumberOfPlanewaves() * HOA_MAXBLKSIZE * sizeof(t_sample));
         ebox_attrprocess_viabinbuf(x, d);
-        
+
         return x;
     }
-    
+
     return NULL;
 }
 
@@ -291,7 +291,7 @@ static void hoa_decoder_3d_perform64(t_hoa_decoder_3d *x, t_object *dsp64, t_sam
     {
         Signal<t_sample>::copy(sampleframes, x->f_outs+i, numouts, outs[i], 1);
     }
-    
+
 }
 
 static void hoa_decoder_3d_dsp(t_hoa_decoder_3d *x, t_object *dsp64, short *count, double samplerate, long maxvectorsize, long flags)
@@ -305,7 +305,7 @@ static t_pd_err hoa_decoder_3d_angles_set(t_hoa_decoder_3d *x, void *attr, long 
     if(argc && argv)
     {
         int dspState = canvas_suspend_dsp();
-        for(long i = 0, j = 0; j < argc && i < x->f_decoder->getNumberOfPlanewaves() * 2; j++)
+        for(long i = 0, j = 0; j < argc && i < (long)x->f_decoder->getNumberOfPlanewaves() * 2; j++)
         {
             if(atom_gettype(argv+j) == A_FLOAT)
             {
@@ -318,7 +318,7 @@ static t_pd_err hoa_decoder_3d_angles_set(t_hoa_decoder_3d *x, void *attr, long 
                 {
                     x->f_decoder->setPlanewaveAzimuth(i, atom_getfloat(argv+j) / 360. * HOA_2PI);
                 }
-                
+
             }
         }
         canvas_resume_dsp(dspState);
@@ -333,7 +333,7 @@ static t_pd_err hoa_decoder_3d_angles_get(t_hoa_decoder_3d *x, void *attr, long 
     *argv = (t_atom *)malloc(*argc * sizeof(t_atom));
     if(*argc && *argv)
     {
-        for(long i = 0; i < x->f_decoder->getNumberOfPlanewaves(); i++)
+        for(ulong i = 0; i < x->f_decoder->getNumberOfPlanewaves(); i++)
         {
             atom_setfloat(*argv+i*2, x->f_decoder->getPlanewaveAzimuth(i) * 360. / HOA_2PI);
             atom_setfloat(*argv+i*2+1, x->f_decoder->getPlanewaveElevation(i) * 360. / HOA_2PI);
@@ -400,25 +400,25 @@ static void hoa_decoder_3d_free(t_hoa_decoder_3d *x)
 extern "C" void setup_hoa0x2e3d0x2edecoder_tilde(void)
 {
     t_eclass* c;
-    
+
     c = eclass_new("hoa.3d.decoder~", (method)hoa_decoder_3d_new, (method)hoa_decoder_3d_free, (short)sizeof(t_hoa_decoder_3d), 0L, A_GIMME, 0);
-    
+
     eclass_dspinit(c);
     hoa_initclass(c);
     eclass_addmethod(c, (method)hoa_decoder_3d_dsp,           "dsp",          A_CANT,  0);
-    
+
     CLASS_ATTR_DOUBLE_VARSIZE	(c, "angles",0, t_hoa_decoder_3d, f_attrs, f_attrs, HOA_MAX_PLANEWAVES*2);
     CLASS_ATTR_ACCESSORS		(c, "angles", hoa_decoder_3d_angles_get, hoa_decoder_3d_angles_set);
     CLASS_ATTR_CATEGORY			(c, "offset", 0, "Planewaves");
     CLASS_ATTR_LABEL			(c, "angles", 0, "Angles of Loudspeakers");
     CLASS_ATTR_SAVE             (c, "angles", 0);
-    
+
     CLASS_ATTR_DOUBLE_ARRAY     (c, "offset", 0, t_hoa_decoder_3d, f_attrs, 3);
     CLASS_ATTR_ACCESSORS		(c, "offset", hoa_decoder_3d_offset_get, hoa_decoder_3d_offset_set);
     CLASS_ATTR_CATEGORY			(c, "offset", 0, "Planewaves");
     CLASS_ATTR_LABEL            (c, "offset", 0, "Offset of Channels");
     CLASS_ATTR_SAVE             (c, "offset", 0);
-    
+
     eclass_register(CLASS_OBJ, c);
     hoa_decoder_3d_class = c;
 }
